requirements.txt
PyJWT==2.7.0
python-dotenv==1.0.0
aiohttp==3.9.1
recnetpy==0.2.71
python-dateutil==2.8.2